match fs_drvsel.c (thanks to cuyler!)

+++ include/JSystem/JUtility/fs_drvsel.h
@@ -8,7 +8,7 @@ extern "C" {
 #endif
 
 extern u16 FS_DrvSel_Reset(void);
-extern u16 FS_DrvSel_Init(int param1, int param2, int param3);
+extern u16 FS_DrvSel_Init(int param1, int param2, u16 param3);
 extern u16 FS_DrvSel_Select(int arg5);
 extern u16 FS_DrvSel_Read(SDSTATUS* arg0, s32 arg1, s32 arg2, s32 arg3, ReadWriteDParam5* arg4, int arg5);
 extern u16 FS_DrvSel_Write(SDSTATUS* arg0, s32 arg1, s32 arg2, s32 arg3, ReadWriteDParam5* arg4, int arg5);
